Remove problematic configure test
The test compiles a file and uses the "strings" program to search for
text in the binary output.

On 2020-06-05, a STRINGS parameter was added to configure that allows the
caller to specify the path to the strings binary. This should be set to a
version of "strings" that can search wasm files.

See:

    https://code.videolan.org/videolan/x264/-/commit/235ce6130168f4deee55c88ecda5ab84d81d125b